09_langgraph_deep_agents/requirements.txt
# Requirements for LangGraph Deep Agents Module
langgraph
langchain
langchain-openai
deepagents
python-dotenv
